requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
python-dotenv
pydantic
duckduckgo-search
crawl4ai
neo4j

pymilvus
rich
typer
imaplib2
email-validator
colorama
tqdm
aiofiles
aiosmtplib
python-dateutil